Move Address to c.s.ethereum.common
