+++ client/src/main/java/com/streamr/client/StreamMessageValidator.java
@@ -5,7 +5,7 @@ import com.streamr.client.protocol.message_layer.AbstractGroupKeyMessage;
 import com.streamr.client.protocol.message_layer.StreamMessage;
 import com.streamr.client.protocol.message_layer.ValidationException;
 import com.streamr.client.rest.Stream;
-import com.streamr.client.utils.Address;
+import com.streamr.ethereum.common.Address;
 import com.streamr.client.utils.AddressValidityUtil;
 import com.streamr.client.stream.KeyExchangeUtil;
 import com.streamr.client.utils.SigningUtil;
