+++ client/src/main/java/com/streamr/client/options/ResendRangeOption.java
@@ -3,7 +3,7 @@ package com.streamr.client.options;
 import com.streamr.client.protocol.common.MessageRef;
 import com.streamr.client.protocol.control_layer.ControlMessage;
 import com.streamr.client.protocol.control_layer.ResendRangeRequest;
-import com.streamr.client.utils.Address;
+import com.streamr.ethereum.common.Address;
 import java.util.Date;
 
 public class ResendRangeOption extends ResendOption {
