+++ client/src/main/java/com/streamr/client/protocol/message_layer/MessageIdAdapter.java
@@ -3,7 +3,7 @@ package com.streamr.client.protocol.message_layer;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
-import com.streamr.client.utils.Address;
+import com.streamr.ethereum.common.Address;
 import java.io.IOException;
 
 public class MessageIdAdapter extends JsonAdapter<MessageId> {
