+++ client/src/main/java/com/streamr/client/rest/Json.java
@@ -2,7 +2,7 @@ package com.streamr.client.rest;
 
 import com.squareup.moshi.Moshi;
 import com.streamr.client.protocol.message_layer.StringOrMillisDateJsonAdapter;
-import com.streamr.client.utils.Address;
+import com.streamr.ethereum.common.Address;
 import java.math.BigDecimal;
 import java.util.Date;
