+++ client/src/main/java/com/streamr/client/rest/StorageNodeInput.java
@@ -1,6 +1,6 @@
 package com.streamr.client.rest;
 
-import com.streamr.client.utils.Address;
+import com.streamr.ethereum.common.Address;
 
 public final class StorageNodeInput {
   private final Address storageNodeAddress;
