+++ client/src/main/java/com/streamr/ethereum/crypto/Sign.java
@@ -1,0 +1,4 @@
+package com.streamr.ethereum.crypto;
+
+public class Sign {
+}
