+++ client/src/test/groovy/com/streamr/client/stream/KeyExchangeUtilSpec.groovy
@@ -8,7 +8,7 @@ import com.streamr.client.protocol.message_layer.GroupKeyResponse
 import com.streamr.client.protocol.message_layer.MessageId
 import com.streamr.client.protocol.message_layer.StreamMessage
 import com.streamr.client.testing.TestingAddresses
-import com.streamr.client.utils.Address
+import com.streamr.ethereum.common.Address
 import com.streamr.client.utils.AddressValidityUtil
 import com.streamr.client.utils.MessageCreationUtil
 import java.nio.charset.StandardCharsets
