+++ client/src/test/groovy/com/streamr/client/subs/CombinedSubscriptionSpec.groovy
@@ -8,7 +8,7 @@ import com.streamr.client.protocol.message_layer.StreamMessage
 import com.streamr.client.testing.TestingAddresses
 import com.streamr.client.testing.TestingContent
 import com.streamr.client.testing.TestingMessageRef
-import com.streamr.client.utils.Address
+import com.streamr.ethereum.common.Address
 import com.streamr.client.stream.GroupKeyStore
 import com.streamr.client.stream.KeyExchangeUtil
 import com.streamr.client.utils.OrderedMsgChain
