+++ client/src/test/groovy/com/streamr/client/utils/OrderedMsgChainSpec.groovy
@@ -6,6 +6,7 @@ import com.streamr.client.protocol.message_layer.StreamMessage
 import com.streamr.client.testing.TestingAddresses
 import com.streamr.client.testing.TestingContent
 import com.streamr.client.testing.TestingMessageRef
+import com.streamr.ethereum.common.Address
 import java.nio.charset.StandardCharsets
 import java.util.function.Consumer
 import java.util.function.Function
