+++ client/src/test/java/com/streamr/client/testing/StreamMessageExamples.java
@@ -3,7 +3,7 @@ package com.streamr.client.testing;
 import com.streamr.client.protocol.common.MessageRef;
 import com.streamr.client.protocol.message_layer.MessageId;
 import com.streamr.client.protocol.message_layer.StreamMessage;
-import com.streamr.client.utils.Address;
+import com.streamr.ethereum.common.Address;
 import java.util.HashMap;
 import java.util.Map;
