+++ client/src/test/java/com/streamr/client/testing/TestingAddresses.java
@@ -1,6 +1,6 @@
 package com.streamr.client.testing;
 
-import com.streamr.client.utils.Address;
+import com.streamr.ethereum.common.Address;
 import java.util.Random;
 
 public final class TestingAddresses {
