+++ client/src/test/java/com/streamr/ethereum/common/AddressTest.java
@@ -1,7 +1,6 @@
-package com.streamr.client.utils;
+package com.streamr.ethereum.common;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import nl.jqno.equalsverifier.EqualsVerifier;
